requirements.txt
langchain
langgraph
langsmith
langchain-groq
psycopg2-binary
pymongo
python-dotenv